Extract parsing to AST logic to a new file out of main to allow reusing within test generator
